test: split test_client
